requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
paho-mqtt==1.6.1
geopy==2.4.1
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.23.2
pyyaml==6.0.1
aiosqlite==0.19.0
